flowcharts
Added flowcharts with DiagrammeR

Used DiagrammeR R commands instead of grViz command for reproducibility. grViz command requires use of DOT language which has a learning curve, while R commands use dataframes.  It seems that being able to use "rank" is not possible in the current iteration of DiagrammeR without the direct and explicit use of DOT/Graphviz so flowchart nodes are not on level planes as in reference images.

Many of the commands on linked DiagrammeR Docs page out of date -- eg  'create_nodes' is now 'create_nodes_df'. Information from that page should be used with caution, but the vignette is up to date.

Additionally, top of slide 15 in French presentation had no text while there was text in the English version. I pasted the English text there as a reminder that this text should be included, since though my translation skills are not sufficient to translate it myself.